Schärfen mit unscharf maskieren hinzugefügt

+++ Project/data/fortran_test/update.py
@@ -3,12 +3,15 @@ import re
 
 # folders = ['FASTEST_1', 'FASTEST_2', 'FASTEST_3', 'FASTEST_4']
 # folders = ['FASTEST_1', 'FASTEST_2', 'FASTEST_3']
-# folders = ['FASTEST_3', 'FASTEST_2']
-# folders = ['FASTEST_1', 'FASTEST_2']
-# folders = ['FASTEST_1']
-folders = ['FASTEST_2']
+# folders = ['FASTEST_1', 'FASTEST_2', 'FASTEST_3']
+folders = ['FASTEST_1']
+# folders = ['FASTEST_2']
 # folders = ['FASTEST_3']
 # folders = ['FASTEST_4']
+
+# folders = ['FASTEST_2', 'FASTEST_3']
+
+# folders = ['FASTEST_1', 'FASTEST_4']
 
 for folder in folders:
     # if folder == 'FASTEST_1':
@@ -19,9 +22,9 @@ for folder in folders:
     print(f'Getting {folder}')
     # Mit res
     # if ('3' in folder) or ('4' in folder):
-    # os.system(f'rsync -vLa --delete-before --delete-excluded --include="res/***" --include="out/***" --include="id/***" --include="map/***" --include="output.txt" --include="funcprp.F" --include="funcusr.F" --include="curv_ml.F" --include="y_pos.txt" --exclude="*" --exclude="out/oscillation.res" --exclude=".map" zkraus@Maira.fnb.maschinenbau.tu-darmstadt.de:/work/local/zkraus/{maira_folder}/projects/02_Oscillation/ /Users/zach/Git/Masterarbeit/Project/data/fortran_test/{folder}/')
+    os.system(f'rsync -vLa --delete-before --delete-excluded --include="res/***" --include="out/***" --include="id/***" --include="map/***" --include="output.txt" --include="funcprp.F" --include="funcusr.F" --include="curv_ml.F" --include="y_pos.txt" --exclude="*" --exclude="out/oscillation.res" --exclude=".map" zkraus@Maira.fnb.maschinenbau.tu-darmstadt.de:/work/local/zkraus/{maira_folder}/projects/02_Oscillation/ /Users/zach/Git/Masterarbeit/Project/data/fortran_test/{folder}/')
     # Aufsteigende Blase
-    # os.system(f'rsync -vLa --delete-before --delete-excluded --include="res/***" --include="out/***" --include="id/***" --include="map/***" --include="funcprp.F" --include="funcusr.F" --include="curv_ml.F" --include="y_pos.txt" --exclude="*" --exclude="out/oscillation.res" --exclude=".map" zkraus@Maira.fnb.maschinenbau.tu-darmstadt.de:/work/local/zkraus/{maira_folder}/projects/rbubble_test1/ /Users/zach/Git/Masterarbeit/Project/data/fortran_test/{folder}/')
+    # os.system(f'rsync -vLa --delete-before --delete-excluded --include="res/***" --include="out/***" --include="id/***" --include="map/***" --include="output.txt" --include="funcprp.F" --include="funcusr.F" --include="curv_ml.F" --include="y_pos.txt" --exclude="*" --exclude="out/oscillation.res" --exclude=".map" zkraus@Maira.fnb.maschinenbau.tu-darmstadt.de:/work/local/zkraus/{maira_folder}/projects/rbubble_test1/ /Users/zach/Git/Masterarbeit/Project/data/fortran_test/{folder}/')
     # Static Bubble mit output.txt
     # os.system(f'rsync -vLa --delete-before --delete-excluded --include="res/***" --include="out/***" --include="id/***" --include="map/***" --include="funcprp.F" --include="funcusr.F" --include="output.txt" --include="curv_ml.F" --include="y_pos.txt" --exclude="*" --exclude="out/oscillation.res" --exclude=".map" zkraus@Maira.fnb.maschinenbau.tu-darmstadt.de:/work/local/zkraus/{maira_folder}/projects/01_StaticBubble/ /Users/zach/Git/Masterarbeit/Project/data/fortran_test/{folder}/')
     # Ohne output.txt
@@ -31,19 +34,20 @@ for folder in folders:
     # if ('2' in folder) or ('1' in folder):
     # os.system(f'rsync -vLa --delete-before --delete-excluded --include="out/***" --include="id/***" --include="map/***" --include="funcprp.F" --include="funcusr.F" --include="curv_ml.F" --include="y_pos.txt" --exclude="*" --exclude=".res" --exclude=".map" zkraus@Maira.fnb.maschinenbau.tu-darmstadt.de:/work/local/zkraus/{maira_folder}/projects/02_Oscillation/ /Users/zach/Git/Masterarbeit/Project/data/fortran_test/{folder}/')
     # Mit output.txt
-    os.system(f'rsync -vLa --delete-before --delete-excluded --exclude=".res" --include="out/***" --include="id/***" --include="output.txt" --include="map/***" --include="funcprp.F" --include="funcusr.F" --include="curv_ml.F" --include="y_pos.txt" --exclude="*" --exclude="out/oscillation.res" --exclude=".map" zkraus@Maira.fnb.maschinenbau.tu-darmstadt.de:/work/local/zkraus/{maira_folder}/projects/02_Oscillation/ /Users/zach/Git/Masterarbeit/Project/data/fortran_test/{folder}/')
+    # os.system(f'rsync -vLa --delete-before --delete-excluded --exclude=".res" --include="out/***" --include="id/***" --include="output.txt" --include="map/***" --include="funcprp.F" --include="funcusr.F" --include="curv_ml.F" --include="y_pos.txt" --exclude="*" --exclude="out/oscillation.res" --exclude=".map" zkraus@Maira.fnb.maschinenbau.tu-darmstadt.de:/work/local/zkraus/{maira_folder}/projects/02_Oscillation/ /Users/zach/Git/Masterarbeit/Project/data/fortran_test/{folder}/')
     # os.system(f'rsync -va --delete-before zkraus@Maira.fnb.maschinenbau.tu-darmstadt.de:/work/local/zkraus/{maira_folder}/projects/02_Oscillation/src/fhp/fsrc/curv_ml.F /Users/zach/Git/Masterarbeit/Project/data/fortran_test/{folder}/')
 
 # '''
 print('Executing regex')
 
 for path in folders:
-    with open('./'+path+'/y_pos.txt', 'r') as myfile:
-        data = myfile.read()
-        data = re.sub(r'(\d) +(\d)', r'\1, \2', data)
-        data = re.sub(r'(?<!,) +', r'', data)
+    if os.path.isfile('./'+path+'/y_pos.txt'):
+        with open('./'+path+'/y_pos.txt', 'r') as myfile:
+            data = myfile.read()
+            data = re.sub(r'(\d) +(\d)', r'\1, \2', data)
+            data = re.sub(r'(?<!,) +', r'', data)
 
-        outfile = open('./'+path+'/y_pos_re.txt', 'w')
-        outfile.write(data)
-        outfile.close()
-# '''
+            outfile = open('./'+path+'/y_pos_re.txt', 'w')
+            outfile.write(data)
+            outfile.close()
+    # '''
